% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hex.R
\name{plot_hex}
\alias{plot_hex}
\title{Hex plot from a data frame}
\usage{
plot_hex(
  df,
  bins = 200,
  hex_cut = c(0.02, 0.98),
  aggr_fun = "mean",
  colorscale = rev(viridis::inferno(256)),
  color_label = "Expression",
  axis_labels = c("UMAP1", "UMAP2")
)
}
\arguments{
\item{df}{Data frame with three columns, the first two are the coordinates
and the third is the expression}

\item{bins}{Number of hex bins}

\item{hex_cut}{Quantiles to use for the color scale}

\item{aggr_fun}{Aggregation function (e.g. mean, median)}

\item{colorscale}{Color scale}

\item{color_label}{Color scale label}

\item{axis_labels}{Axis labels}
}
\value{
A ggplot2 object
}
\description{
Plot a hex bin plot from a data frame with some default
parameters for a nice looking plot, when not using a Seurat object.
}
