% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hex.R
\name{plot_hex_minimal}
\alias{plot_hex_minimal}
\title{Minimal hex plot from a data frame}
\usage{
plot_hex_minimal(
  df,
  bins = 200,
  hex_cut = c(0.02, 0.98),
  aggr_fun = "mean",
  colorscale = rev(viridis::inferno(256))
)
}
\arguments{
\item{df}{Data frame with three columns, the first two are the coordinates
and the third is the expression}

\item{bins}{Number of hex bins}

\item{hex_cut}{Quantiles to use for the color scale}

\item{aggr_fun}{Aggregation function (e.g. mean, median)}

\item{colorscale}{Color scale}
}
\value{
A ggplot2 object
}
\description{
Plot a hex bin plot from a data frame. Good starting point for
customizing the plot.
}
