% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hex.R
\name{HexPlot}
\alias{HexPlot}
\title{HexPlot}
\usage{
HexPlot(
  srt,
  features,
  dims = c(1, 2),
  aggr.fun = "mean",
  color.scale = rev(viridis::inferno(256)),
  bins = 200,
  min.cutoff = NA,
  max.cutoff = NA,
  reduction = "umap",
  split.by = NULL,
  slot = "data",
  coord.fixed = FALSE,
  combine = TRUE
)
}
\arguments{
\item{srt}{Seurat object}

\item{features}{Feature(s) to plot. Either a single feature or a vector of
features}

\item{dims}{Dimensions to plot}

\item{aggr.fun}{Aggregation function (e.g. mean, median)}

\item{color.scale}{Color scale}

\item{bins}{Number of hex bins}

\item{min.cutoff}{Minimum expression cutoff}

\item{max.cutoff}{Maximum expression cutoff}

\item{reduction}{Which dimensionality reduction to use}

\item{split.by}{A factor in object metadata to split the feature plot by,
pass 'ident' to split by cell identity'. Used by Seurat::SplitObject}

\item{slot}{Which slot to use for the feature expression}

\item{coord.fixed}{Whether to use a fixed aspect ratio}

\item{combine}{Whether to combine the plots into a single plot using
patchwork or return a list of plots. Only applies when split.by is not NULL
or features is a vector of length > 1}
}
\value{
A ggplot2 object if split.by is NULL, otherwise a patchwork object
}
\description{
Plot a hex bin plot of a feature in a Seurat object
}
\examples{
library(Seurat)
srt <- pbmc_small
HexPlot(srt, "CD3D", reduction = "tsne")
}
